Edge weights should be doubles

+++ Minimum-Spanning-Tree/Edge.h
@@ -5,14 +5,14 @@
 class Edge
 {
 public:
-	Edge(int edgeWeight, int startVertexId, int endVertexId)
+	Edge(double edgeWeight, int startVertexId, int endVertexId)
 	{
 		weight = edgeWeight;
 		startingVertexId = startVertexId;
 		endingVertexId = endVertexId;
 	};
 	~Edge();
-	int weight;
+	double weight;
 	unsigned int startingVertexId;
 	unsigned int endingVertexId;
 };
